requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
torch>=2.0.0
torchvision>=0.15.0
open-clip-torch>=2.20.0
faiss-cpu>=1.7.0
numpy>=1.24.0
Pillow>=10.0.0
requests>=2.31.0
python-multipart>=0.0.6
tqdm>=4.66.0
python-dotenv>=1.0.0
pydantic>=2.0.0
